sachen

+++ MeinBurger/src/meinburger/Bun.java
@@ -2,10 +2,10 @@ package meinburger;
 
 public class Bun extends Ingredients{
 	
-	public int backzeit ; 
+	public double backzeit ; 
 	public int hoehe;
 
-	public Bun(int number, String name, float price, boolean classic, boolean vegan, boolean vegetarian, int hoehe, int backzeit) {
+	public Bun(int number, String name, float price, boolean classic, boolean vegan, boolean vegetarian, int hoehe, double backzeit) {
 		super(number, name, price, classic, vegan, vegetarian);
 		this.hoehe = hoehe; 
 		this.backzeit = backzeit;
